+++ CHANGELOG.rst
@@ -10,7 +10,7 @@ Change Log
 =====
 `PR 27 SN Add RIN required validator <https://github.com/smaht-dac/submitr/pull/27>`_
 
-* Add a validator that ensures that the property rna_integrity_number has a value for Analyte items where "RNA" is in molecule
+* Add a validator that ensures that the property rna_integrity_number has a value for Analyte items where "RNA" is in molecule and non-RNA analytes cannot have a value for `rna_integrity_number`
 
 
 1.8.0